hook up user auth
