+++ src/assets/LoginPage.css
@@ -7,7 +7,7 @@
   width: auto;
 }  
   
-  .create-account-btn {
+  .login-account-btn {
     margin-top: 1.5rem;
   }
