+++ src/model/user/IUserAuthRequest.tsx
@@ -1,0 +1,4 @@
+export default interface IUserAuthRequest {
+    username: string,
+    password: string
+}
