+++ src/model/user/IUserAuthResponse.tsx
@@ -1,0 +1,5 @@
+export default interface IUserAuthResponse {
+    token: string,
+    userId: number,
+    username: string
+}
